LayoutTests/fast/selectors/querySelector-mixed-child-adjacent-backtracking-expected.txt

Test backtracking of adjacent and child relation mixed together.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("html>head+body div>div~h1+blockquote>p+div+blockquote~ul p").length is 8
PASS document.querySelectorAll("html>head+body div>div~h1+blockquote>p+div+blockquote~ul p")[0].id is "target1"
PASS document.querySelectorAll("html>head+body div>div~h1+blockquote>p+div+blockquote~ul p")[1].id is "target2"
PASS document.querySelectorAll("html>head+body div>div~h1+blockquote>p+div+blockquote~ul p")[2].id is "target3"
PASS document.querySelectorAll("html>head+body div>div~h1+blockquote>p+div+blockquote~ul p")[3].id is "target4"
PASS document.querySelectorAll("html>head+body div>div~h1+blockquote>p+div+blockquote~ul p")[4].id is "target5"
PASS document.querySelectorAll("html>head+body div>div~h1+blockquote>p+div+blockquote~ul p")[5].id is "target6"
PASS document.querySelectorAll("html>head+body div>div~h1+blockquote>p+div+blockquote~ul p")[6].id is "target7"
PASS document.querySelectorAll("html>head+body div>div~h1+blockquote>p+div+blockquote~ul p")[7].id is "target8"
PASS successfullyParsed is true

TEST COMPLETE
